Feature/add training session factory and seeders

+++ src/seeders/user.seeder.ts
@@ -1,8 +1,9 @@
 import { Connection } from 'mongoose'
-import { createAdminUser, createSuperAdminUser, createUser } from '../factories/user.factory'
+import { createAdminUser, createSuperAdminUser, createUsers } from '../factories/user.factory'
 import userService from '../services/user.service'
 import logger from '../utils/logger'
 import { CreateUserDTO } from '../DTOs/user/create.dto'
+import { checkCollectionExistence } from '../utils/database.utils'
 
 const USER_COUNT = 20
 
@@ -11,7 +12,7 @@ async function seedUsers(db: Connection) {
     await deleteUsers(db)
     await Promise.all([userService.create(createSuperAdminUser()), userService.create(createAdminUser())])
 
-    const users: CreateUserDTO[] = Array.from({ length: USER_COUNT }, () => createUser())
+    const users: CreateUserDTO[] = createUsers(USER_COUNT)
     await Promise.all(users.map((user) => userService.create(user)))
 
     logger.info('Users created successfully')
@@ -21,11 +22,14 @@ async function seedUsers(db: Connection) {
 }
 
 async function deleteUsers(db: Connection) {
-  try {
-    await db.collection('users').drop()
-    logger.info('Users collection has been dropped')
-  } catch (error) {
-    logger.error(error)
+  const collectionExists = await checkCollectionExistence(db, 'users')
+  if (collectionExists) {
+    try {
+      await db.collection('users').drop()
+      logger.info('Users collection has been dropped')
+    } catch (error) {
+      logger.error(error)
+    }
   }
 }
